Simplify data fetching
