+++ src/Api/filterables.json
@@ -1,0 +1,69 @@
+[
+  {
+    "id": "type",
+    "field": "type",
+    "display": "Type",
+    "options": ["proposal", "publication"],
+    "group": "documents",
+    "type": "list"
+  },
+  {
+    "id": "remote-list",
+    "type": "remote-list",
+    "id": "technique",
+    "field": "name",
+    "display": "Technique",
+    "group": "techniques",
+    "optionsUrl": "https://techniques.scicat.ess.eu/techniques?filter=%7B%22where%22%3A%7B%22and%22%3A%5B%7B%22pid%22%3A%7B%22nlike%22%3A%22%3Adataset%22%7D%7D%2C%7B%22name%22%3A%7B%22nlike%22%3A%20%22CHMO%22%7D%7D%5D%7D%2C%22fields%22%3A%7B%22pid%22%3Atrue%2C%22name%22%3Atrue%7D%7D"
+  },
+  {
+    "id": "incident_wavelength",
+    "field": "incident_wavelength",
+    "display": "Incident Wavelength",
+    "type": "numeric",
+    "group": "parameters",
+    "units": ["nm"]
+  },
+  {
+    "id": "incident_photon_energy",
+    "field": "incident_photon_energy",
+    "display": "Incident Photon Energy",
+    "type": "numeric",
+    "group": "parameters",
+    "units": ["eV", "J"]
+  },
+  {
+    "id": "sample_chemical_formula",
+    "field": "sample_chemical_formula",
+    "display": "Chemical Formula",
+    "group": "parameters",
+    "operator": "like"
+  },
+  {
+    "id": "sample_temperature",
+    "field": "sample_temperature",
+    "display": "Temperature",
+    "type": "numeric",
+    "group": "parameters",
+    "units": ["K", "C", "F"]
+  },
+  {
+    "id": "sample_pressure",
+    "field": "sample_pressure",
+    "display": "Pressure",
+    "type": "numeric",
+    "group": "parameters",
+    "units": ["Pa"]
+  },
+  {
+    "id": "count_events",
+    "field": "count_events",
+    "display": "count_events",
+    "group": "parameters",
+    "type": "numeric"
+  },
+  {
+    "id": "limit",
+    "field": "limit"
+  }
+]
