+++ src/Api/translate.test.js
@@ -1,35 +1,28 @@
-import { template } from '../../filters'
-import { LABEL_FOR_CONFIG } from './lib/state'
-import applyTemplate from './translate'
-
-const translate = applyTemplate(template)
-
-const id = LABEL_FOR_CONFIG
+import translate from './translate'
 
 test('no filter and default config', () => {
   const query = translate([])
-  expect(query).toEqual({ limit: 5 })
+  expect(query).toEqual({ limit: 50 })
 })
 
 test('no pagination', () => {
-  const query = translate([{ id, limit: '0' }])
-  expect(query).toEqual({ limit: '0' })
+  const query = translate([{ limit: 0 }])
+  expect(query).toEqual({ limit: 0 })
 })
 
 test('custom ordering', () => {
   const query = translate([
     {
-      id,
       order: ['foo ASC', 'bar DESC'],
     },
   ])
 
-  expect(query).toEqual({ order: ['foo ASC', 'bar DESC'], limit: 5 })
+  expect(query).toEqual({ order: ['foo ASC', 'bar DESC'], limit: 50 })
 })
 
 test('single root filter and custom include', () => {
   const query = translate([
-    { id, include: ['datasets', 'affiliation', 'person'] },
+    { include: ['datasets', 'affiliation', 'person'] },
     {
       id: 'type',
       value: 'proposal',
@@ -47,13 +40,13 @@ test('single root filter and custom include', () => {
       { relation: 'datasets' },
     ],
     where: { type: 'proposal' },
-    limit: 5,
+    limit: 50,
   })
 })
 
 test('multiple filters and custom include', () => {
   const query = translate([
-    { id, include: ['datasets', 'affiliation', 'person'] },
+    { include: ['datasets', 'affiliation', 'person'] },
     {
       id: 'type',
       value: 'experiment',
@@ -95,7 +88,7 @@ test('multiple filters and custom include', () => {
       },
     ],
     where: { type: 'experiment' },
-    limit: 5,
+    limit: 50,
   })
 })
 
@@ -104,7 +97,6 @@ test('all loopback query keys recognized by id', () => {
     { id: 'skip', value: 20 },
     { id: 'limit', value: 20 },
     { id: 'include', value: ['parameters'] },
-    { id: 'fields', value: ['pid', 'title', 'description'] },
     { id: 'order', value: ['foo ASC', 'bar DESC'] },
   ])
   expect(query).toEqual({
@@ -117,6 +109,5 @@ test('all loopback query keys recognized by id', () => {
     order: ['foo ASC', 'bar DESC'],
     skip: 20,
     limit: 20,
-    fields: { pid: true, title: true, description: true },
   })
 })
