+++ src/App/helpers.js
@@ -8,3 +8,9 @@ export function parseDate(date) {
 export function capitalizeAndSpace(str) {
   return `${str[0].toUpperCase()}${str.slice(1).replaceAll('_', ' ')}`
 }
+
+const isEmpty = (obj) => JSON.stringify(obj) === '{}'
+export const stripEmptyKeys = (obj) =>
+  Object.fromEntries(
+    Object.entries(obj).filter(([, v]) => v !== undefined && !isEmpty(v)),
+  )
