+++ src/Explore/DocumentList.js
@@ -1,33 +1,31 @@
 import React, { useEffect } from 'react'
-import useSWRImmutable from 'swr/immutable'
 
+import useApi from '../Api/useApi'
+import useFilters from '../Api/useFilters'
 import { useSearchStore } from '../App/stores'
 import { Flex, Card, Text, Heading } from '../Primitives'
 import DocumentItem from './DocumentItem'
 
-function DocumentList(props) {
-  const { queryUrl } = props
+function DocumentList() {
   const setCount = useSearchStore((state) => state.setCount)
-
-  const { data: documents } = useSWRImmutable(queryUrl)
+  const filters = useFilters()
+  const { data } = useApi('/documents', filters)
 
   useEffect(() => {
-    setCount(documents.length)
-  }, [documents, setCount])
+    setCount(data.length)
+  }, [data, setCount])
 
   return (
     <Flex column gap={[3, 3, 3, 4]}>
-      {documents.length === 0 ? (
+      {data.length === 0 ? (
         <Card p={[3, 4]}>
           <Heading>No results</Heading>
           <Text as="p">Please adjust the search filters.</Text>
         </Card>
       ) : (
-        <>
-          {documents.map((doc) => (
-            <DocumentItem document={doc} key={doc.pid} />
-          ))}
-        </>
+        data.map((document) => (
+          <DocumentItem document={document} key={document.pid} />
+        ))
       )}
     </Flex>
   )
