--- src/filterables.json
@@ -1,50 +1,0 @@
-{
-  "documents": [
-    {
-      "field": "type",
-      "display": "Type",
-      "options": ["proposal", "publication"],
-      "type": "list"
-    }
-  ],
-  "techniques": [
-    {
-      "type": "remote-list",
-      "id": "technique",
-      "field": "name",
-      "display": "Technique",
-      "optionsUrl": "https://techniques.scicat.ess.eu/techniques?filter=%7B%22where%22%3A%7B%22and%22%3A%5B%7B%22pid%22%3A%7B%22nlike%22%3A%22%3Adataset%22%7D%7D%2C%7B%22name%22%3A%7B%22nlike%22%3A%20%22CHMO%22%7D%7D%5D%7D%2C%22fields%22%3A%7B%22pid%22%3Atrue%2C%22name%22%3Atrue%7D%7D"
-    }
-  ],
-  "parameters": [
-    {
-      "field": "incident_wavelength",
-      "display": "Incident Wavelength",
-      "type": "numeric",
-      "units": ["nm"]
-    },
-    {
-      "field": "incident_photon_energy",
-      "display": "Incident Photon Energy",
-      "type": "numeric",
-      "units": ["eV", "J"]
-    },
-    {
-      "field": "sample_chemical_formula",
-      "display": "Chemical Formula",
-      "operator": "like"
-    },
-    {
-      "field": "sample_temperature",
-      "display": "Temperature",
-      "type": "numeric",
-      "units": ["K", "C", "F"]
-    },
-    {
-      "field": "sample_pressure",
-      "display": "Pressure",
-      "type": "numeric",
-      "units": ["Pa"]
-    }
-  ]
-}
